requirements.txt
Pillow==9.4.0
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
Flask==2.2.2